delete_model.py: added option --delete-only-versions

+++ mlflow_tools/tools/delete_model.py
@@ -8,7 +8,7 @@ import mlflow
 print("MLflow Tracking URI:", mlflow.get_tracking_uri())
 
 
-def delete_model(model_name, client=None):
+def delete_model(model_name, delete_only_versions=False, client=None):
     """ Delete a model and all its versions. """
     if not client:
         client = mlflow.client.MlflowClient()
@@ -19,15 +19,17 @@ def delete_model(model_name, client=None):
         if vr.current_stage != "Archived": # NOTE: for Databricks though OSS works
             client.transition_model_version_stage (model_name, vr.version, "Archived")
         client.delete_model_version(model_name, vr.version)
-    client.delete_registered_model(model_name)
+    if not delete_only_versions:
+        client.delete_registered_model(model_name)
 
 
 @click.command()
 @click.option("--model", help="Registered model name", required=True, type=str)
-def main(model):
+@click.option("--delete-only-versions", help="Delete only versions and not the registered model", default=False, type=bool)
+def main(model, delete_only_versions):
     print("Options:")
     for k,v in locals().items(): print(f"  {k}: {v}")
-    delete_model(model)
+    delete_model(model, delete_only_versions)
 
 
 if __name__ == "__main__":
